Added profile page components
Each component of the profile page.

profilepage.html has the tabs of each section of the page and the
picture and info of the user.

tab-wishlist.htlm has the view of the wishlist tab.

tab-album.html has the view of the album tab.

tab-calender.html has the view of the calender tab.

Also, CSS of the profile components.